+++ src/os2datascanner/projects/report/reportapp/views/report_views.py
@@ -44,6 +44,7 @@ from os2datascanner.engine2.rules.dict_lookup import EmailHeaderRule
 from os2datascanner.engine2.rules.passport import PassportRule
 
 from .utilities.ews_utilities import try_ews_delete
+from .utilities.google_utilities import try_gmail_delete
 from .utilities.smb_utilities import try_smb_delete_1
 from .utilities.document_report_utilities import handle_report
 from .utilities.msgraph_utilities import delete_email, delete_file
@@ -132,6 +133,9 @@ class ReportView(LoginRequiredMixin, ListView):
         context["show_ews_delete_button"] = settings.EWS_ALLOW_WRITE
         context["show_ews_mass_delete_button"] = settings.EWS_ALLOW_WRITE and \
             self.all_reports_from_same_source("ews", context["page_obj"])
+        context["show_gmail_delete_button"] = settings.GMAIL_ALLOW_WRITE
+        context["show_gmail_mass_delete_button"] = settings.GMAIL_ALLOW_WRITE and \
+            self.all_reports_from_same_source("gmail", context["page_obj"])
 
         # Retention policy details
         context["retention_policy"] = self.org.retention_policy
@@ -366,6 +370,7 @@ class UndistributedView(PermissionRequiredMixin, ReportView):
         context["show_smb_mass_delete_button"] = False
         context["show_ews_delete_button"] = False
         context["show_ews_mass_delete_button"] = False
+        context["show_gmail_delete_button"] = False
         return context
 
 
@@ -388,6 +393,7 @@ class ArchiveMixin:
         context["show_smb_mass_delete_button"] = False
         context["show_ews_delete_button"] = False
         context["show_ews_mass_delete_button"] = False
+        context["show_gmail_delete_button"] = False
         return context
 
     def dispatch(self, request, *args, **kwargs):
@@ -713,6 +719,50 @@ class MassDeleteEWSMailView(HTMXEndpointView, BaseMassView):
 
     def delete_ews_mails(self, document_reports):
         deleted, problem = try_ews_delete(
+            self.request, document_reports.values_list(
+                "pk", flat=True))
+
+        if not deleted:
+            error_message = _("Failed to delete some reports: {e}").format(
+                e=problem)
+            messages.add_message(
+                self.request,
+                messages.WARNING,
+                error_message)
+
+
+class DeleteGmailView(HTMXEndpointView, DetailView):
+    """ View for sending a delete request for a gmail."""
+    model = DocumentReport
+
+    def post(self, request, *args, **kwargs):
+        response = super().post(request, *args, **kwargs)
+        report = self.get_object()
+
+        deleted, problem = try_gmail_delete(request, [report.pk])
+        if not deleted:
+            error_message = _("Failed to delete {pn}: {e}").format(
+                pn=report.matches.handle.presentation_name, e=problem)
+            messages.add_message(
+                request,
+                messages.WARNING,
+                error_message)
+
+        return response
+
+
+class MassDeleteGmailView(HTMXEndpointView, BaseMassView):
+    """View for sending delete requests for multiple Gmails."""
+
+    def post(self, request, *args, **kwargs):
+        response = super().post(request, *args, **kwargs)
+        reports = self.get_queryset()
+        self.delete_gmails(reports)
+
+        return response
+
+    def delete_gmails(self, document_reports):
+        deleted, problem = try_gmail_delete(
             self.request, document_reports.values_list(
                 "pk", flat=True))