+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - Google Drive results now also has a button that opens the associated file.
 
+- It is now possible to delete Gmail emails from OSdatascanner. 
+
 ### General improvements
 
 - Adjusted responsive size of navigation buttons in the admin module.